pyproject.toml
[tool.poetry]
name = "coce"
version = "0.3.0"
description = "Cluster Outlier Catalyst Engine"
authors = ["Benji"]
packages = [{include = "src"}]
python = "^3.10"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2"
numpy = "^1.26"
scipy = "^1.12"
networkx = "^3.3"
python-louvain = "^0.16"
scikit-learn = "^1.5"
statsmodels = "^0.14"
ccxt = "^4.2"
textblob = "^0.17"
pyyaml = "^6.0"
python-dotenv = "^1.0"
tqdm = "^4.66"
joblib = "^1.4"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
ruff = "^0.4"
mypy = "^1.10"
pytest = "^8.0"
pytest-cov = "^5.0"
types-requests = "^2.31"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
src = ["src"]
line-length = 100
